fix payment

+++ database/migrations/2025_05_21_132726_create_food_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->timestamps();
             $table->string('name');
             $table->text('description')->nullable();
-            $table->decimal('price', 8, 2);
+            $table->integer('price');
             $table->string('image_url')->nullable();
         });
     }
